+++ dpAutoRigSystem/Modules/Library/jcRibbon.py
@@ -371,7 +371,7 @@ class RibbonClass(object):
             cmds.addAttr(curve, longName='autoRotate', attributeType='float', minValue=0, maxValue=1, defaultValue=0.2, keyable=True)
         else:
             cmds.addAttr(curve, longName='autoRotate', attributeType='float', minValue=0, maxValue=1, defaultValue=0, keyable=True)
-        cmds.addAttr(curve, longName='pin', attributeType='float', minValue=0, maxValue=1, defaultValue=0, keyable=True)
+        cmds.addAttr(curve, longName='pin', attributeType='long', minValue=0, maxValue=1, defaultValue=0, keyable=True)
         self.dpUIinst.ctrls.setLockHide([curve], ['v'])
         return [grp, curve, zero0, zero1]
